[*] add animator component editor support

+++ Sources/Overload/OvRendering/include/OvRendering/Resources/Animation.h
@@ -22,7 +22,7 @@ namespace OvRendering::Resources
         ~Animation();
         BoneFrames* FindBone(const std::string& name);
         
-        inline float GetTicksPerSecond() { return m_TicksPerSecond; }
+        inline int GetTicksPerSecond() { return m_TicksPerSecond; }
         inline float GetDuration() { return m_Duration; }  
 
     public:
